analysis/pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=2.8.1",
    "openai-whisper>=20250625",
    "pyannote-audio>=4.0.1",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "whisper>=1.1.10",
]
